# harness/web/src/services/cde/swagger.yaml
openapi: 3.0.0
info:
  title: API Specification
  version: 0.0.0
servers:
- url: gateway/cde/api/v1
security:
- bearerAuth: []
  x-api-key: []
paths:
...
  /accounts/{accountIdentifier}/orgs/{orgIdentifier}/projects/{projectIdentifier}/gitspaces/{gitspace_identifier}/token:
    get:
      operationId: getToken
      parameters:
      - description: account identifier.
        in: path
        name: accountIdentifier
        required: true
        schema:
          type: string
      - description: org identifier.
        in: path
        name: orgIdentifier
        required: true
        schema:
          type: string
      - description: project identifier.
        in: path
        name: projectIdentifier
        required: true
        schema:
          type: string
      - description: gitspace identifier.
        in: path
        name: gitspace_identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenapiGetTokenResponse'
          description: OK
      summary: Get gitspace token
      tags:
      - gitspaces
...
components:
  schemas:
...
    OpenapiGetTokenResponse:
      properties:
        gitspace_token:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
    x-api-key:
      in: header
      name: x-api-key
      type: apiKey